requirements.txt
# Core Dependencies
python-dotenv>=1.0.0          # Environment variable management

# HTTP and Web
requests>=2.31.0              # HTTP requests
beautifulsoup4>=4.12.0        # HTML parsing
lxml>=4.9.0                   # XML/HTML parser
selenium>=4.15.0              # Browser automation

# Google Workspace APIs
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# Data Processing
pandas>=2.1.0                 # Data manipulation
openpyxl>=3.1.0              # Excel file handling
numpy>=1.26.0                # Numerical operations

# AI/ML APIs
openai>=1.6.0                # OpenAI API client
anthropic>=0.8.0             # Anthropic API client

# Utilities
pytz>=2023.3                 # Timezone handling
python-dateutil>=2.8.2       # Date utilities
jsonschema>=4.20.0           # JSON validation
pyyaml>=6.0.1                # YAML parsing

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging
colorlog>=6.8.0              # Colored logging output

# MCP Server
mcp[cli]>=1.0.0              # Model Context Protocol server SDK
